Filter Metadata

filter_metadata allows users to retreive a data frame of variables based on a set of filter values for variable categories, or the entire set of metadata as a data frame if no filters are provided

filter_metadata(filter_list = list(), ...)

Arguments

filter_list
a named list of variables to filter metadata on. See details for valid field names.

...
additional value combinations to filter on. See details for valid field names.

Value

returns data frame of variable metadata based on a set of given filter values, or the entire set of metadata if no filters are provided

Details

List of valid field names:

new_name: the new, standardized variable name

old_name: the original variable name from older files

type: variable type. Options are binary (bin), continuous (cont), unordered categorical (uc), ordered categorical (oc), and string (string)

warning: flags for issues

group: group number (matches the same questions across surveys)

q_group_N: number of variables that are in the group

topic1: indicates specific content topic (e.g. medication or housing status), there is no hierarchy between topic1 and topic2

topic2: indicates specific content topic (e.g. medication or housing status), there is no hierarchy between topic1 and topic2

source: source of variable

respondent: respondent, either the person interviewed or the place an interview took place

wave: the wave of data collection associated with the variable

scope: the reach of a variable's sample population (grouped by city where each person was originally interviewed). The max scope is 20 cities.

section: section of survey where question is asked

leaf: the rest of the variable name

q_group_list: a string that lists all of the new variable names for all the variables in the group

value1-value78: all of the potential responses for a given variable

label1-label68: all of the labels for the potential responses for a given variable

Examples

filter_test <- filter_metadata(wave = 3, source = "constructed", type = "bin")