Select Metadata

select_metadata allows users to retrieve the value of a specific field of a given variable, or the entire variable if no field is specified

select_metadata(variable_name = NULL, field_name = NULL)

Arguments

variable_name
name of variable

field_name
specific field of variable to be accessed. See details for valid field names

Value

returns string with value of a given field if field is specified, returns a data.frame with variables for every field if field is unspecified

Details

List of valid field names:

new_name: the new, standardized variable name

old_name: the original variable name from older files

type: variable type. Options are binary (bin), continuous (cont), unordered categorical (uc), ordered categorical (oc), and string (string)

warning: flags for issues

group: group number (matches the same questions across surveys)

q_group_N: number of variables that are in the group

topic1: indicates specific content topic (e.g. medication or housing status), there is no hierarchy between topic1 and topic2

topic2: indicates specific content topic (e.g. medication or housing status), there is no hierarchy between topic1 and topic2

source: source of variable

respondent: respondent, either the person interviewed or the place an interview took place

wave: the wave of data collection associated with the variable

scope: the reach of a variable's sample population (grouped by city where each person was originally interviewed). The max scope is 20 cities.

section: section of survey where question is asked

leaf: the rest of the variable name

q_group_list: a string that lists all of the new variable names for all the variables in the group

value1-value78: all of the potential responses for a given variable

label1-label68: all of the labels for the potential responses for a given variable

Examples

select1 <- select_metadata(variable_name = "ce3agefc")
select2 <- select_metadata(variable_name = "ce3agefc", field_name = "type")